+++ src/core/lib/event_engine/posix_engine/posix_engine.h
@@ -222,7 +222,7 @@ class PosixEventEngine final : public EventEngine,
       const grpc_event_engine::experimental::PosixTcpOptions& options,
       Duration timeout);
 
-  void OnConnectFinishInternal(int connection_handle);
+  void OnConnectFinishInternal(int64_t connection_handle);
 
   std::vector<ConnectionShard> connection_shards_;
   std::atomic<int64_t> last_connection_id_{1};
